Clean up some things and fix bullet spawning issue.
Codemaid clutching up making the scripts look pretty.
Also did some general organization cleaning in the editor.
The bullet spawning issue was that a bullet would spawn, but it would instantly disappear. Commented out the OnEnable function in the Bullet script and it fixed the issue. My guess is the bullet thought it hit the maxDistance before it actually did.